Card, Player and Store Fixes
Changed the Card to be a class specific cards can inherit all of their
actions from, implemented images in the cards.  Fixed small issues in
the Player (player was starting with three Villages instead of 3
estates) and in the store (an infinite amount of cards were being added
to the store stacks).  One thing of note: Right now the store stack
arrays store the same card in all of their slots.  This doesn't really
matter much, as currently the player actions only check against there
being a card in the stack, then remove a card from the stack and create
one out of thin air for themselves, but it might come up later.  Making
the Card cloneable may help.