More Turn Logic and PlayerPanel updates
First off, PlayerPanel has been turned into an abstract class that
extends JPanel, and what was PlayerPanel is now ActionPanel, the panel
that lets you use cards.  The only method in the PlayerPanel abstract
class so far is update, though I'm sure we'll need to add more.  We have
the abstract PlayerPanel class so we can store different types of panels
in the same PlayerPanel item in the view.

I've also added a little bit more turn logic to both the GameManager and
the Player classes, and we shouldn't have too much trouble with turns
from here on in.  We'll need to control the turn order using the button
commands we create in the different PlayerPanel subclasses.

Everythings still a little confusing, but this should work out alright.
We'll need to pass the View (the main master class that runs and
displays everything) to anything that does turn stuff, including the
PlayerPanels and the card Actions.