View Update working
View now has an update method that can be called to switch out the
controls based on the status of the player.  The different Player Panels
and Card Actions will set the player status and then update the view for
each type of phase, whether through normal phase turns or through
special actions.